requirements.txt
django
django-rest-framework
psycopg2-binary
django-cors-headers
djangorestframework
typing-extensions
redis
pillow
